Mod downloading.

+++ src/main/java/sk/tomsik68/mclauncher/api/ui/IProgressMonitor.java
@@ -13,23 +13,23 @@ public interface IProgressMonitor {
      * You need to scale it yourself in order to get progress information between 0 and 100.</B>
      *
      * @param progress New progress value
-     * @see IProgressMonitor#setMax(int)
+     * @see IProgressMonitor#setMax(long)
      */
-    public void setProgress(int progress);
+    public void setProgress(long progress);
 
     /**
      * Sets maximum progress value
      *
      * @param len Max progress value
      */
-    public void setMax(int len);
+    public void setMax(long len);
 
     /**
      * Increments current progress value by <code>amount</code>
      *
      * @param amount How much
      */
-    public void incrementProgress(int amount);
+    public void incrementProgress(long amount);
 
 
     /**
